Moved lifecycle reducers into their own files for clarity.
