+++ server/src/lifecycle/client_disconnected.rs
@@ -1,0 +1,19 @@
+use spacetimedb::*;
+use spacetimedsl::*;
+
+use crate::tables::global_config::*;
+
+#[spacetimedb::reducer(client_disconnected)]
+pub fn identity_disconnected(ctx: &ReducerContext) -> Result<(), String> {
+    let dsl = dsl(ctx);
+    // Called everytime a client disconnects
+
+    if let Some(mut config) = dsl.get_all_global_configurations().next() {
+        if *config.get_active_players() > 0 {
+            config.set_active_players(config.get_active_players() - 1);
+            dsl.update_global_config_by_id(config)?;
+        }
+    }
+
+    Ok(())
+}
