+++ server/src/lifecycle/mod.rs
@@ -1,0 +1,3 @@
+pub mod client_connected;
+pub mod client_disconnected;
+pub mod init;
